+++ Sources/VComponents/Components/Buttons/Link/VLink.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 // MARK: - V Link
-/// Button component that controls a navigation presentation to an URL.
+/// Button component that controls a navigation presentation to an `URL`.
 ///
 /// Button is meant to be used with button components from the library.
 ///
@@ -17,12 +17,7 @@ import SwiftUI
 ///     var body: some View {
 ///         VLink(
 ///             url: .init(string: "https://www.apple.com")!,
-///             label: {
-///                 VSecondaryButton(
-///                 action: {},
-///                 title: "Lorem Ipsum"
-///                 )
-///             }
+///             title: "Lorem Ipsum"
 ///         )
 ///     }
 ///
@@ -31,26 +26,47 @@ public struct VLink<Label>: View where Label: View {
     @Environment(\.isEnabled) private var isEnabled: Bool
     
     private let url: URL
-    private let label: () -> Label
+    private let label: VLinkLabel<Label>
 
     // MARK: Initializers
-    /// Initializes component with url and label.
+    /// Initializes component with `URL` and title.
+    public init(
+        url: URL,
+        title: String
+    )
+        where Label == Never
+    {
+        self.url = url
+        self.label = .title(title: title)
+    }
+    
+    /// Initializes component with `URL` and label.
     public init(
         url: URL,
         @ViewBuilder label: @escaping () -> Label
     ) {
         self.url = url
-        self.label = label
+        self.label = .custom(label: label)
     }
 
     // MARK: Body
     public var body: some View {
         Link(
             destination: url,
-            label: label
+            label: buttonLabel
         )
             .buttonStyle(.plain) // Cancels styling
             .disabled(!isEnabled)
+    }
+    
+    @ViewBuilder private func buttonLabel() -> some View {
+        switch label {
+        case .title(let title):
+            VPlainButton(action: {}, title: title)
+            
+        case .custom(let label):
+            label()
+        }
     }
 }
 
@@ -59,12 +75,7 @@ struct VLink_Previews: PreviewProvider {
     static var previews: some View {
         VLink(
             url: .init(string: "https://www.apple.com")!, // fatalError
-            label: {
-                VSecondaryButton(
-                    action: {},
-                    title: "Lorem Ipsum"
-                )
-            }
+            title: "Lorem Ipsum"
         )
     }
 }