+++ Sources/VComponents/Components/Buttons/Link/VLinkLabel.swift
@@ -1,0 +1,11 @@
+//
+//  VLinkLabel.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 03.07.22.
+//
+
+import Foundation
+
+// MARK: - V Link Label
+typealias VLinkLabel = GenericLabel_TitleCustom