+++ Sources/VComponents/Components/Buttons/Navigation Link/VNavigationLinkLabel.swift
@@ -1,0 +1,11 @@
+//
+//  VNavigationLinkLabel.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 03.07.22.
+//
+
+import Foundation
+
+// MARK: - V Navigation Link Label
+typealias VNavigationLinkLabel = GenericLabel_TitleCustom